Rewrite query for latest features
The old query used one subquery that calculated max timestamp in a
range. The new query introduces an additional subquery that limits the
joined data to the range we want, and the join is performed on the
results of those two subqueries.

This is more efficient in presence of a metaindex on the timestamp
column, because it may significantly reduces the number of objects that
need to be read.

Signed-off-by: Michal Siedlaczek <michal.siedlaczek@ibm.com>